backend/requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
click==8.3.2
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.11
MarkupSafe==3.0.3
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
